userportal: Fix exception on browser resizing
The SideTabExtendedResourceView listens to Window.addResizeHandler
and tries to change it's widget hights.

The problem is, that this recieves the events also when the resources
view is not visible so the widgets' heights are 0 leaving the
calculated new height negative. This throws an assertion exception:
"CSS heights should not be negative".

Fixed by guarding that this size will never be less than zero.

Change-Id: I43238b399a7e4148ebb3c75b3afdac807ad78202
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>