gluster: Add host: Improved gluster validation [2]
Currently we don't allow to add a host if it has peers. However, if one
(or more) of those peers are already part of the cluster in engine DB,
then it means that these have been added from gluster CLI and hence the
validation should not fail in such cases.

Change-Id: I36b3b6d00a71af50ef1b006e95daebfbfe3f67f5
Signed-off-by: Shireesh Anjal <sanjal@redhat.com>