core: save memory state on live snapshot with memory
On creation of live snapshot with memory, we create two volumes, one for
the memory dump and one for the vm configuration at the moment of the
snapshot creation (that vdsm saves).

Those volumes are stored on the snapshot entity in a similar way to how
the hibernation volumes which are created on suspend VM command are
stored, and they are passed to the live snapshot verb in vdsm.

Change-Id: Ic7ceb07bf73e722bf4566989446e7dbba78a2ccb
Bug-Url: https://bugzilla.redhat.com/960931
Signed-off-by: Arik Hadas <ahadas@redhat.com>