engine: Exception handling - VDSCommandBase
The following patch will simplify exception handling in case of failure in xml rpc call.
All XmlRpcRunTimeException will be translated to network exception with error code VdcBllErrors.VDS_NETWORK_ERROR.
The XmlRpcRunTimeException - means that we failed in call , failed to parse xml and we did not get appropriate response
that we can parse, this is by definition VdcBllErrors.VDS_NETWORK_ERROR.
Now, VdcBllErrors.VDS_PROTOCOL error was removed as unused.

Change-Id: I0bd38cd75706ff7669c1ff055b951aeb864b2736
Signed-off-by: Michael Kublin <mkublin@redhat.com>