core: Remove StoragePoolIsoMapDAO.update impl
Since the owner column was removed from StoragePoolIsoMap in the
previous patch, it's only (database) mutable column is the status
column.
In other words, there is no difference between the update and
updateStatus methods.

This patch removes the Updatestorage_pool_iso_map stored procedure
and replaces the implementation of the update method with a
delegation to updateStatus.

Change-Id: Ia661353ed7b100795d90c9390866c955a1a6290d
Signed-off-by: Allon Mureinik <amureini@redhat.com>