pki: move configuration from vdc_options to configuration files
root application serves pki resources, in order to support variable
location we need to be able to access these without accessing the
database.

in addition, there is no point in storing file locations within
database, as if we use cluster we have no grantee to have these at the
other node.

handling the root application is not included in this patch.

Change-Id: I1764d9ca7a8c677401f721b3d89f45deff9c1f26
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>