core, webadmin: Move VmValidationUtils
This class has no business being in common - it's conceptually part of
the bll validation framework, and is indeed only used in bll.

Note:
As part of the change, its entry was removed from Common.gwt.xml (as
bll classes cannot be imported to GWT), with no adverse effect.
This only strengthens the point that this class was not supposed to be
in common.

Change-Id: I506dc65f8740cdc18d57007f46a75d7f89f169b1
Signed-off-by: Allon Mureinik <amureini@redhat.com>