frontend: Spice is always available
This patch removes the method ConsoleUtils.isSpiceAvailable because it is no
longer needed as spice can be always invoked using the nonplugin way.

This patch also moves configuration of spice implementations (by Configurator
classes) and connect listener to the method responsible for switching
implementations on the fly so that effect of these actions is not forgotten
after switching spice implementation.

Change-Id: Ice5c3f80988962c5fa5e39f51e91543938258912
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>