core: Fix testGetRootDSENoServers
Since urls is a zero length list, urls.get(0) throws an
IndexOutOfBoundsException.
This exception is caught, and thus the test passes, but it's kind of
meaningless - it does not execute any of the DirectorySearcher's code,
and essentially tests Java's ArrayList.get(int) implementation.

I guess it's safe to assume that it works correctly.

Change-Id: I295d9b10495efabf892107187de7ccd91334062f
Signed-off-by: Allon Mureinik <amureini@redhat.com>