core: Add log when task could not end action.
Adding a log when a task try to end an action but encounter a scenario
that not all the tasks related to the entity were ended also.

Without this log, it will be hard to realize that the entity has more
tasks in it, if there was a bug that the tasks were not cleaned up
properly.
An example for this bug can be seen in the end of the log of the related bug
BZ958061.
The ended tasks are logged there over and over again but do not end the
command properly.

Change-Id: Ia1e4700676622b23c79a388b6aa3b92ae1aa15e9
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>