core: move image group command
Added MoveImageGroupCommand - making MoveOrCopyImageGroup command to be
used for copy only (CopyImageGroup).

Reasons for that change:
1. When moving an image and attempting to delete with wipe - the
operation will be faster rather then waiting for the wipe to complete.
2. Moving towards lowering the SPM involvement in the copy/move process.
3. The engine will be able to tell on which part of the "move" operation
an error has occured.
4. Move operation on vdsm should be deprecated.

Known issues and reasoning:
* The whole operation might be slower for moving image groups without wipe
as the delete part is quick (and now we will have two vdsm calls rather
then one) - this issue is known and accepted as part of this RFE.

Change-Id: Id9068d66df3986c9bb16b266bb5bef396964a706
Bug-Url: https://bugzilla.redhat.com/753549
Signed-off-by: Liron Aravot <laravot@redhat.com>