webadmin: [ja] Edit pool- prestarted vms string is broken into two lines
In 'ja' locale the text of prestarted vms is too long for 230px.
Changed all the right column widgets to be 250px.

Also changed the width of the popup to 710px (Changing the width of the
rigth widgets caused "Increase number of VMs in pool by" to be broken
into two lines. BTW, in some langauges it was broken even before the
change).

The change applies only on EditPool.

Change-Id: Iaf34f3adbde08bb34d8db7db3508ba3cf4c31f92
Bug-Url: https://bugzilla.redhat.com/950788
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>