core: reduce compensation scope from hibernate command
This patch reduce the scope in which the compensation that is defined
to switch the VM status back to UP is set in the execute stage of the
hibernate vm command.

The compensation context is cleared when the first task is created (in
transaction) - there's no need to keep the compensation anymore since we
don't want the status of the VM to switch to UP while tasks that were
created as part of the hibernate command exist.

Note: this is only part of the solution for the bug mentioned below.

Change-Id: Iec46798fc6ea01ef6b0b4701e1af5ba616edf910
Bug-Url: https://bugzilla.redhat.com/949281
Signed-off-by: Arik Hadas <ahadas@redhat.com>