core: introduce VM status PrepareForHibernate
This patch adds status for VM that reflects the state in which the VM is
going to be hibernated but vdsm didn't start to save the memory state
yet.

When starting the hibernate command, this is the first status to be set
fot the vm, then the images which are going to be used are created, and
only when the 'hibernate' verb in the vdsm is called and vdsm reports
back that it starts to do the migration to the file, the VM status will
be changed to SavingState.

Change-Id: Iebc8e0a7cd4b21e751393304ceee65d668d84ca4
Bug-Url: https://bugzilla.redhat.com/949281
Signed-off-by: Arik Hadas <ahadas@redhat.com>