core: lock the vds manager while calling hibernate verb
It is needed to lock the vds manager so that no statuses will be updated
while calling the hibernate verb in vdsm and setting the status of the
VM to SavingState afterwards.

The previous HibernateVDSCommand is changed to
HibernateBrokerVDSCommand, and a new HibernateVDSCommand that calls it
after the vds manager is locked, is added.

Change-Id: I2537fffccd15cb5fa30f99793c803373886c633f
Bug-Url: https://bugzilla.redhat.com/949281
Signed-off-by: Arik Hadas <ahadas@redhat.com>