packaging: engine-service: protect configuration files
we need to protect configuration files as database password resides
within.

for some reason jboss changes the permissions of configuration file and
deployment to 0777, enforcing umask of 0027 perform some remedy at the
price of having logs not world readable as well, so it is not a
solution.

luckily, jboss does not modify the directory in which configuration file
resides, so we split the configuration directory to own directory and
set proper permissions.

Change-Id: Ie76e5ffcdccd8c3a0bedbb9ce9bd56e146eeccd9
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>