core: [cleanup] introduce GetAllFromExportDomainQuery class
This patch changes hierarchy which is logically wrong and reduces the
amount of code duplication:

Before this patch GetTemplatesFromExportDomainQuery inherited
GetVmsFromExportDomainQuery which is wrong since there's no kind-of
relationship between those classes, and there was duplicated code
between those two classes.

GetAllFromExportDomainQuery is added as a parent for the above classes -
they both are kind-of query that returns all from export domain thus the
relations are logically ok, and the new class contains code which is
join for both classes.

This hierarchy will make it easier to add query for instance-types from
export domain later.

Change-Id: I2544cb879898ef90d64feb8a16dcfb8cf103b409
Signed-off-by: Arik Hadas <ahadas@redhat.com>