core: add missing sound card for imported vm
This patch solves a bug in which when importing desktop vm that doesn't
have sound card device (which is the case when importing vm that was
created in 3.0 for example), the resulting vm won't have sound device.

The solution is to check the imported vm and if it is desktop vm and it
doesn't have sound device, we add sound device to it.

Change-Id: I295d158f014f865be5a03f630ce48ba446752e5e
Bug-Url: https://bugzilla.redhat.com/955498
Signed-off-by: Arik Hadas <ahadas@redhat.com>