gluster: set service name&type in ServerServiceDAO
Introduced new view gluster_server_services_view with additional columns
for service_name and service_type.

Updated SPs to fetch from this view instead of the table. (Removed one
duplicate SP in this process)

Modified GlusterServerServiceRowMapper in the DAO to populate service
name and type. Also updated the test case to verify this.

Change-Id: I6a4706aae779e73b4aae7ca5810fe998559f62a4
Signed-off-by: Shireesh Anjal <anjalshireesh@gmail.com>