usertportal: Console buttons are still active after VM is launched.
After spice console connecting/disconnecting in extended usertportar view, the
ui state of the console buttons was not updated. This behavior is reproducible
since patch 09f1acafa2a508f90ea9a7a77f79693d8e8224d6 in upstream.

The problem was in not comparing console state in the logic that was
responsible for deciding whether to refresh the view. Fixed by adding
check for console state change (console username).

Change-Id: I6b260c338dbc2f1156f00732842caf59d7fd1bb2
Bug-Url: https://bugzilla.redhat.com/951666
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>