restapi: Attaching a network to cluster by network id
The server throws an internal server error when
add network command is executed with invalid network
name or when providing an id only.

The patch adds support to attaching a network to a cluster
either by network ID or network name or both, as long as
there is a match between the ID to the name of the network.

Bug-Url: https://bugzilla.redhat.com/961046
Change-Id: Ie5fac9af37fec642fd3c0c6fbdb43ee04b364f2c
Signed-off-by: Moti Asayag <masayag@redhat.com>