engine: Avoid lower-case l
Using a lower case "l" to denote a long literal is visually confusing
since an "l" is graphically similar to a "1" in a lot of fonts.
Thus, using an upper-case "L" is considered less confusing.

This patch fixes all the occasions where a lower case "l" is used and
adds a checkstyle validation to prevent this mistake from
re-occurring.

Change-Id: I33cafbcc46e4162d7f31871a16c52a511d4c90c4
Signed-off-by: Allon Mureinik <amureini@redhat.com>