core: Compensate to unknown status to avoid rollback to active status.
On deactivate storage domain command, the storage domain(s) status is being
saved in the compensation context so in case of failure - it will
be reverted to its original value.

In case of failure when executing DeactiveStorageDomainVDSCommand this failure
will propagate to DeactiveStorageDomainCommand and will cause to the compensation
saved data to be applied - which might override the changed statuses and return the
domain status to be active again.

In this change, the domain status is being saved in the compensation
context in unknown status rather then active, so in case of failure the
status would be reverted to unknown.

Change-Id: I9f8a27cbac711fa87db408a06fe22a7b3e9b2da9
Bug-Url: https://bugzilla.redhat.com/920694
Signed-off-by: Liron Aravot <laravot@redhat.com>