core: Fix unauthorized move-disk
When quota was set to 'Enforcing' user could move a disk to a storage
domain which he had no permissions to use.

For some reason, an 'if' statement was used in order to add the
storage-domain to the permission subject list only when quota was set to
'Disabled' or 'Audit'. 'If' statement was canceled. The storage-domain
is added to the list when disk is about to be moved to the
storage-domain.

Change-Id: I32713cbcf3206ab449ddca9b0ebd382bf77a7582
Signed-off-by: Ofri Masad <omasad@redhat.com>