userportal,webadmin: type ahead list box
A new component ListModelTypeAheadListBoxEditor has been created.

It looks like a regular combo box, but the behavior is as follows:
- when the user clicks inside or to the left triangle (combo box)
  all the options are shown (like for regular combo box)
- it supports type ahead
- the display options are rendered using custom renderers and
  can even contain HTML
- it is bound to the model

It is based on the code from ListModelSuggestBox. The common parts
have been moved to BaseListModelSuggestBox and generalized. The
behavior of the original ListModelSuggestBox(Editor) is unchanged.

The new widget has been used on the AbstractVmPopupWidget for:
- data center
- cluster
- quota
- template

Change-Id: I956af3c675894c850a1a104a81cec49f4bd62011
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>