Validate installed host by requesting for vdsm uuid
Deploying an host requires to verify if an host has vdsm installed.
The command 'vdsm-tool vdsm-id' returns vdsm uuid if vdsm exists and
installed on host and it doesn't require running vdsm instance.
During canDoAction of addVdsCommand we check this variable and compare
it to vds table. If exists we fail the operation before starting deploy.

Need to have http://gerrit.ovirt.org/#/c/14866 in vdsm side to work
properly.

Change-Id: I5c6c0a85daba47fabb9253963ff187a670f28ae6
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>