gluster: Command to update gluster hook on servers
This patch adds a command to update hook content
on servers where the content's different.
Either the engine content or content from a server
is used - depending on parameters -
to update content on servers.
If content from server is used, the engine's copy
of hook content is updated as well.

Also, refactored the GlusterHookParameters class
and GlusterHookCommandBase class

Change-Id: Ibccccbbed1e07d471dc472c41843f46767b9e083
Signed-off-by: Sahina Bose <sabose@redhat.com>