userportal,webadmin: unify new server/desktop to newVm
From now there is no new Server and new Desktop - only one new VM.
The server/desktop distinction is done by a new combo box:
Optimized for: Server/Desktop

From now there are no options disabled for server or desktop, just
a set of defaults which changes according to the selection in the
combo box.

This defaults only includes:
- soundcard is by default enabled for desktop and disabled for server
- provisioning: thin for desktop, clone for server

This defaults can be overridden.

A new checkbox has been introduced to the Console tab for the VM:
Soundcard enabled (previously this has been determined by vm type,
now it has to be explicitly set).

If new vm is pressed, the default settins are server.

For new pool the default is desktop.

The vm type can be changed in both new and edit dialogs.

Change-Id: Ibcb7065e2968a84fef30951e8a45019ea1095520
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>