packaging: setup: support package upgrade
yum reads exclude/include information to sack when first used. this has
a potential conflict with the overall transaction as the ordering
determine when the versionlock is read.

otopi was modified to create a new miniyum instance at when transaction
is started. this enables the PROGRAMS stage to run and read a fresh copy
of versionlock.

we cannot handle the versionlock using the regular transaction as it its
unlocked content should wrap the transaction or we have random yum
incompatibilities.

also, we cannot retrieve our package version in order to perform the
lock, as the setup package may be newer than the other packages, so use
rpm directly in order to figure out the specific versions of packages.

Change-Id: I399fb48d8cbd393ee12f700c0a68e7c1b51218f3
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>