core: Manage OS properties in a repository
This is the implementation of the OS info according to
http://www.ovirt.org/OS_info

This feature enables ovirt to configure all its OS info in one place,
out of the code, and in a way admins could extend it or easily configure
it.
It aims to merge all configuration items from the Config (currently
vdc_options table) and the VmOsType.java enum to one file-based
repository.

This patch contains:

* The default repo config file osinfo-defaults.properties
* A loading logic to support multiple properties file loading
* A Repository which holds an inner tree with all the configuration
* A backend query to expose the data to UI

Change-Id: I50384bf3b6763fa7a34badc2449e06242186a858
Signed-off-by: Roy Golan <rgolan@redhat.com>