restapi: 'async' URL Parameter Malfunction (#957452)
If 'async' param appears in URL, then, regardless of the value assigned to it (true/false),
the application would, erroniously, always consider it true. After the change, 'async=true', 'async' are considered true,
'async=false' or the absence of this parameter are considered false, and an illegal value ('async=blabla') throws exception.

Change-Id: I8b8ac5d49823612300f1c7872c8d317a20e02797
Bug-Url: http://bugzilla.redhat.com/957452
Signed-off-by: Ori Liel <oliel@redhat.com>