engine: Integrate SPICE-HTML5 client
This patch integrates SPICE-HTML5 client in the engine.

It makes use of the websocket proxy integrated in I44e9870b.

Invocation by this client can be enabled in the console popup. For this
to be enabled, the websocket proxy must be set in the engine.

Moreover this patch adds parameter ssl_target to the ticket which is
passed to the proxy. This parameter is necessary for telling the proxy
that the (SPICE) endpoint uses ssl. The format of the ticket has changed
from positional style to encoded json.

Change-Id: I490cb57f5d23a7a763807d3bee00ea0e4c032338
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>