engine: Unify Guid and NGuid
Guid and NGuid are essentially the same class, where the only
difference is the value returned from createGuidFromString(String) in
case the parameter is null (null vs. Guid.Empty) and a couple of
constants.

Since this method is static anyway, and most places use Guid anyway,
NGuid can be removed as a separate class.
Moreover, NGuid's hierarchy is obviously broken, as it holds a cache for
a Guid instance, which in turn, extends NGuid.

This patch does the following:
1. Renames Guid.createGuidFromString(String) to
   createGuidFromStringDefaultEmpty, in order to overload it with
   NGuid.createGuidFromString(String).
2. Squashed NGuid and Guid together and retained the name Guid. All
   occurrences of NGuid in the code were replaced with Guid.
3. Renamed NGuidTest to GuidTest, since NGuid no longer exists.
4. Renamed JsonNGuidMixIn to JsonGuidMixIn, since NGuid no longer
   exists.
5. Removed specific references in JsonObjectSerializer,
   JsonObjectDeserializer and their tests.
6. Removed NGuid's entry from Compat.gwt.xml.

This patch leaves some work to be done, which was not including simply
because of the size/scope of this patch.
Consecutive patches will handle these cleanups.

Change-Id: I4af6d353ef27394ff902f374de19682628c52739
Signed-off-by: Allon Mureinik <amureini@redhat.com>