gluster: Bll command to refresh gluster hooks in engine
Used to manually trigger a sync of gluster hooks in
gluster cluster with values stored in engine database.
This is needed because the default sync interval is
set to 1 hour as hooks are not expected to change
frequently. If admin wishes to see updated status,
can trigger refresh of hooks from UI.

Change-Id: I4713ca800432aee302f913d0c8aa06a35a5b2db2
Signed-off-by: Sahina Bose <sabose@redhat.com>