core: RegisterDiskCommand - correct storage domain
When registering new disk, the disk was saved to the db using the domain
information stored on its metadata - this can cause to failure during
registration or to have incorrect information stored in the engine db.
The domain on which the disk was existed on isn't necessarily the
new one on which the disk is now, so it shouldn't be saved with it.

The following patch set the saved storage domain to the current one and
is part of the complete fix which involves also setting the current id
to the image metadata in vdsm.

Change-Id: I568668fc1a8024ff255754287afbaa5d35b8a2ed
Signed-off-by: Liron Aravot <laravot@redhat.com>