restapi: memory snapshot feature support
This patch adds support in the REST API for the memory snapshot feature
which is about saving the memory state as part of snapshots, so that it
will be possible to restore the same state of the VM as it was at the
moment the snapshot was taken, when the VM is reverted to the snapshot.

An attribute of type boolean which indicates whether the snapshot
contains memory or not is added to the snapshot entity.

The user can select when creating a snapshot whether the created
snapshot should include memory state or not.

The user can select when previewing and restoring snapshots whether
to restore the snapshot's memory (which means that the memory that is
saved in the snapshot, if exists, will be restored when running the VM
for the first time after the action is done) or not.

Change-Id: I9ccbbaf9a4d242255a4ff999e5f08644d45402eb
Bug-Url: https://bugzilla.redhat.com/960931
Signed-off-by: Arik Hadas <ahadas@redhat.com>