webadmin: avoid multiple failovers caused by import subtabs auto-refresh
The SubTabStorageTemplateBackupView/SubTabStorageVmBackupView subtabs
executes an engine query to display the vms/templates that exist on the
selected export domain - this engine query executes an spm command
(GetVmsInfoVDSCommand) which will attempt to perform failover in case of
failure.
The fact that the query is being executed every 5 seconds, causes to the
requests to "pile up" and possibly perform multiple failovers, Therefore
the automatic refresh to the tab has been replaced with manual refresh button.

Change-Id: I2668a60b5b33b2ca8d82b9d08c477ba7c1a3a723
Signed-off-by: Liron Aravot <laravot@redhat.com>