core: Power Management failed if no proxy ...
core: Power Management failed if no proxy preferences is defined for the host

Power Management failed if no pm_proxy preferences is defined for the
host.
The pm_proxy_preferences is a per-host field in vds_static that defines
how to look for a proxy host. (cluster then DC or vice versa)
In case that pm_proxy preference is not defined , it should be taken
from the FenceProxyDefaultPreferences configuration value which is
defaulted to "cluster,dc"
However, the code that checks for existence of pm_proxy preferences
checks for null only and if it is an empty string , it is used
The error on the log is :

 2013-06-02 13:04:25,013 ERROR [org.ovirt.engine.core.bll.FenceExecutor]
 Illegal value in PM Proxy Preferences string , skipped

Change-Id: Id1b96b115dc586e6ebbdacd44e0a034e91609841
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=969778
Signed-off-by: Eli Mesika <emesika@redhat.com>