packaging: Running systemctl reload after service update
In 3.2 installations we did not handle %postun section correctly.
It means that the systemd daemon is not reloaded when 3.2 package
is upgraded. This leads to an 'unstartable' ovirt-engine service
when we upgrade to 3.3. As such, during the upgrade to 3.3 we need to
reload the systemd daemon 'manually'.

This patch runs %systemd_postun macro in a %post section
of the spec file during an upgrade flow only.

As soon as we move to post-3.3 releases, this will no longer
be necessary.

Change-Id: If36d81fe89218fb3078b281263df5c89907d900f
Signed-off-by: Alex Lourie <alourie@redhat.com>