engine: Deprecate new Guid()
Guid's default constructor created a new Guid object equal to
Guid.Empty, which is quite redundant since Guids are immutable.
Moreover, the default constructor just delegated to the String
constructor, forcing an expensive UUID.fromString(EMPTY_GUID_VALUE)
to be called on each invocation of new Guid().

In order to solve the performance issues here, the following was done:
1. The default constructor was changed to call the cheap UUID
   constructor instead of the expensive String constructor.
2. The default constructor was changed to private and marked as
   @Deprecated so it won't be called externally. Unfortunately, it could
   not be removed completely due to GWT limitations.
3. All the usages of new Guid() we removed and replaced by Guid.Empty
   for production code or by Guid.newGuid() for testing code (in order
   to provide some randomness, and hence, better robustness for the
   test).

Change-Id: Id80dc50896a9874731300f543a72838c345cbcdd
Signed-off-by: Allon Mureinik <amureini@redhat.com>