restapi: a potential NPE when a user invokes actions
When a user invokes an action without sending filter=true property,
the action may throw a NPE on command creation flow.

The underlining issue is in AbstractBackendCollectionResource
-> resolveCreated method. If an exception is thrown (e.g. due to
lack of permissions to execute query), the method returns null.

Instead of tolerating the exception, call handleError method
(to output a proper error message).

Change-Id: I55f5ae906d6c0d2023783e7b61abe3bf9d01f28f
Bug-Url: https://bugzilla.redhat.com/915106
Signed-off-by: Daniel Erez <derez@redhat.com>