restapi: adding custom properties to vm interface
In Device custom properties feature we are adding custom properties to a
vnic.

It works in the same way custom properties of vm work.
When getting the vnic it will be displayed in the following way-
<custom_properties>
   <custom_property name="prop_name" value="prop_value"/>
</custom_properties>
When adding/updating a vnic the custom properties are parsed to a
string and set on VmNetworkInterface.

Change-Id: Ie7f5daae4b0b2117d0cadceb2bb292fabf445c3c
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>