webadmin: add gateway to edit network from setup networks dialog
When editing an IP configuration, we need to enable the gateway field for
all networks, not only ovirtmgmt.
If the user chose static IP, then the gateway field should be added and
be editable. If DHCP was chosen then the gateway field should be visible
and display the gateway configured via DHCP.

Change-Id: Ie87f8fa3827b216483b553a8e36011382930585c
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>