gluster: SP to update gluster server service
Introduced a stored procedure to update the gluster server service by
server id and service id. The same is used in
ManageGlusterServiceCommand to update the details of the server services
post start/stop/restart.

Already existing stored procedure for update needs id field to be
available for update to work and same is not available during BLL
command execution. So the simple update stored procedure cannot be used
as is for the purpose of updating the server service status.

Change-Id: I2518516c93a264e4a7107ded6f24c3e338f03e4e
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>