16. core : Handle incomplete tasks on server restart
This patch changes the way asyc task placeholders are
cleared on server restart for commands that are partially
submitted to vdsm.

If a task has 4 placeholders and only 2 have been submitted
to vdsm, the database has 2 tasks with vdsm id and 2 with
empty vdsm id on server restarted.

When the server is restarted the tasks are grouped by root
command id and the commands that have partially submitted
tasks are handled differently. The task placeholder with
empty vdsm id are removed from the database. The tasks
which have vdsm id are added to the partiallyCompletedCommandTasks
list. On storage up event when AddStoragePoolExistingTasks is called
The task is marked for failure and once polling is done, they are
failed and removed from database.

http://www.ovirt.org/Features/AsyncTaskManagerChanges_3.3
Change-Id: I3a979d816d9ec8cf25119a33742c8e6af4ff42a8
Signed-off-by: Ravi Nori <rnori@redhat.com>