frontend,engine: inherit template permissions on VM creation
A new checkbox "Copy Permissions" has been introduced
into the new VM dialog and make template dialog.

For new VM dialog:
It is visible only when a non blank template is selected and only
in advanced mode.
If it is checked, all the template permissions (only the direct ones
not the inherited) are copied to the VM.

For make tamplate dialog:
Always visible
If checked all permissions of the VM are copied to the template.

Change-Id: I6e6fddcea410ab8fee5851468980b8eeb0dad428
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>