core: Reduce host connection timeout
This change is the second part of 3c1329. There two new parameters
were introduced to control the host connection timeout and the number
of retries when a connection fails with an IO error. The values
introduced on that change were selected so that they don't change the
behaviour of the system. This change updates the default values so
that hosts will be considered failed and thus fenced in a reasonable
amount of time, thus making the high availability functionality more
predictable. The host connection timeout goes from 180 seconds to 2
seconds, and the number of retries goes from 3 to 0 (no retries).

Change-Id: I52c703607e0ba52b1059f9a774a507debeaa3a6e
Bug-Url: https://bugzilla.redhat.com/863211
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>