core: Add snapshot validation for empty guid.
Currently when virt-v2v is used to convert a vm from a foreign
hypervisor (Xen, VMware and etc), it sets the vm_snapshot_id of
all disks of that vm to empty guid.
This later causes issues doing different tasks on snapshots.

The proposed fix is to add a validation whether the
snapshot id is empty and provides an appropriate message.

Change-Id: I19bc2982adecca198f8dc38ca011f4528dd58db9
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>